chore: clean up decorative section-header comments
Removes or simplifies section-header comments that carry no navigation
value, while keeping labels where a reader genuinely benefits from
seeing sections called out. Specific cases:

- Strip 3-line box decoration (``// ----``/``// Label``/``// ----``)
  to plain ``// Label`` comments in ``automated_reasoning.rs``,
  ``s3/service/mod.rs``, ``elasticache/service.rs``,
  ``conformance-macros/lib.rs``, and e2e test files for bedrock/s3/elasticache.
- Delete per-function numbered labels in ``automated_reasoning_workflows.rs``
  — each section was one ``pub fn`` whose name already carries the label.
- Delete single-function section headers in ``user_pools.rs``
  (``Custom Attributes``, ``Signing Certificate``) — one function per
  section makes the header pure decoration.
- Delete stale/misleading section headers discovered during review:
  ``OpsItems`` above association code in ``ssm/service/associations.rs``,
  ``Document extras`` above ``get_ops_summary`` in ``ssm/service/ops.rs``,
  ``Associations`` above ``update_patch_baseline`` in
  ``ssm/service/patches.rs``, and the orphaned ``Stubs`` header at the
  end of the impl block in ``ssm/service/documents.rs``.
- Extract Resource Servers out of ``cognito/service/identity_providers.rs``
  into a new ``cognito/service/resource_servers.rs`` — the section was
  216 lines of code unrelated to the file name, a natural module split.

No behavior change.
