+++ crates/fakecloud-cognito/src/service/user_pools.rs
@@ -777,8 +777,6 @@ impl CognitoService {
         Ok(AwsResponse::ok_json(response))
     }
 
-    // ── Custom Attributes ──────────────────────────────────────────────
-
     pub(super) fn add_custom_attributes(
         &self,
         req: &AwsRequest,
@@ -962,8 +960,6 @@ impl CognitoService {
             "ClientSecrets": secrets
         })))
     }
-
-    // ── Signing Certificate ────────────────────────────────────────────
 
     pub(super) fn get_signing_certificate(
         &self,
