+++ crates/fakecloud-ssm/src/service/documents.rs
@@ -784,9 +784,7 @@ impl SsmService {
         Ok(AwsResponse::ok_json(json!({})))
     }
 
-    // -----------------------------------------------------------------------
     // Resource policies
-    // -----------------------------------------------------------------------
 
     pub(super) fn put_resource_policy(
         &self,
@@ -919,10 +917,6 @@ impl SsmService {
             )),
         }
     }
-
-    // -----------------------------------------------------------------------
-    // Stubs
-    // -----------------------------------------------------------------------
 }
 
 /// Convert document content between JSON and YAML formats.
